Run input trim outside of thread
It's probably also possible to store only a single reference using an
Arc but all of the iterations that I try then have the compiler
complaining about not knowing the size of the [u8] ahead of time when
dereferencing it back into its inner value. I also tried various ways of
passing the byte slice directly using an Arc but that also gave me too
many compiler errors. The input is so small that I don't see it making a
large difference anyway, so this is probably the cleanest that this
solution will be.